+++ components/date-picker/components/Panel.jsx
@@ -8,7 +8,7 @@ import TimePanel from './TimePanel'
 import { getView, genNewDates } from '../utils'
 
 const Panel = () => {
-  const { outDate, type, onPick, localeDatas, showTime, theme, weekOffset, locale } = useContext(DPContext)
+  const { outDate, type, onPick, localeDatas, showTime, theme, weekOffset, locale, onSelect } = useContext(DPContext)
   const [view, setView] = useState(getView(type))
 
   const [calRenderDates, setCalRenderDates] = useState([])
@@ -22,6 +22,7 @@ const Panel = () => {
   }, [outDate])
 
   const onCalenderPick = (date) => {
+    onSelect(date, true)
     if (type === 'year' || (type === 'month' && view === 'month')) {
       // year || month picker
       onPick(