+++ components/date-picker/index.d.ts
@@ -20,6 +20,7 @@ export interface CommonProps {
   placeholder?: string | string[]
   format?: string
   onChange?: (date: Date | DateRange, dateStr: string | DateRangeString) => void
+  onSelect?: (date: Date, isCompleted: boolean) => void
 }
 type Shortcuts = {
   title: string